This update significantly improves scraper performance, by as much as 6.5x. The scraper component has been split out into a AWS Elastic Container Service (ECS) docker service and a trigger lambda function. The docker image used for the scheduled spider now also includes the scraper service script. This service sits idle until there are case numbers in the scraper queue, in which case it processes them with a default concurrency of 10. Other changes include: - Removes DynamoDB and SNS components. - Separates scheduled spider trigger and invocation by 5 minutes due to race condition. - Moves ECS cluster definition and execution role from spider stack to static stack. - Adds ignore_errors flag for Spider class so the ECS task doesn’t shit the bed when it hits a random DNS resolution exception.
NOTE: To deploy this update, you’ll first need to delete the Scraper and Docker-Repo cloud formation stacks.

+++ src/spider/scheduled_spider.py
@@ -5,12 +5,15 @@ from mjcs.config import config
 from mjcs.util import fetch_from_queue
 
 if __name__ == '__main__':
-    item = fetch_from_queue(config.spider_queue, 1)[0]
+    item = fetch_from_queue(config.spider_queue, 1)
+    if not item:
+        raise Exception('No messages in spider queue')
+    item = item[0]
     params = json.loads(item.body)
     today = datetime.now().date()
     end_date = today - timedelta(days=params['days_ago_start'])
     start_date = today - timedelta(days=params['days_ago_end'])
-    spider = Spider(quiet=True)
+    spider = Spider(quiet=True, ignore_errors=True)
     print(f'Spidering for cases filed between {start_date} and {end_date}')
     item.delete()
     spider.search(start_date=start_date, end_date=end_date)
